requirements.txt
SpeechRecognition
pyttsx3
pyaudio
psutil
pywin32
requests
Pillow
